+++ plugins/ticks/scratch.js
@@ -1,6 +1,6 @@
-const spacetimeTicks = require('./src')
+import spacetimeTicks from './src/index.js'
 
-console.time('d')
+console.time('time')
 let ticks = spacetimeTicks('jan 1 2019', 'jan 1 2020', 12)
 console.log(ticks)
-console.timeEnd('d')
+console.timeEnd('time')